improve mobile locations table (#756)

+++ frontend/src/ui/views/events/edit/tabs/TabLocations.vue
@@ -1,7 +1,7 @@
 <template>
     <div class="full-width-scrollable">
         <VTable
-            :items="props.event.locations"
+            :items="props.event?.locations"
             class="scrollbar-invisible interactive-table no-header"
             :class="$attrs.class"
             :sortable="hasPermission(Permission.WRITE_EVENT_DETAILS)"
@@ -12,24 +12,58 @@
                 <td :key="item?.icon" class="pr-4 text-xl">
                     <i class="fa-solid" :class="item?.icon" />
                 </td>
-                <td class="w-full max-w-[50vw] sm:w-1/2 sm:max-w-full">
+                <td class="w-full sm:w-2/3 lg:w-1/2">
                     <p class="font-semibold">
                         <span>{{ item?.name }}</span>
                     </p>
-                    <p v-if="item?.address" class="mt-1 truncate text-sm font-light">
-                        <span>{{ item.address }}</span>
+                    <div class="flex opacity-75">
+                        <p class="mt-1 w-0 grow truncate text-sm font-light">
+                            <span v-if="item?.address">
+                                {{ item.address }}
+                            </span>
+                            <template v-if="item?.eta || item?.etd">
+                                <span v-if="item.address" class="sm:hidden"> / </span>
+                                <span v-if="item.eta" class="sm:hidden">
+                                    {{ $t('domain.location.eta') }}:
+                                    {{ $d(item.eta, DateTimeFormat.DDD_DD_MM_hh_mm) }}
+                                </span>
+                                <span v-if="item.eta && item.etd" class="sm:hidden"> / </span>
+                                <span v-if="item.etd" class="sm:hidden">
+                                    {{ $t('domain.location.etd') }}:
+                                    {{ $d(item.etd, DateTimeFormat.DDD_DD_MM_hh_mm) }}
+                                </span>
+                            </template>
+                        </p>
+                    </div>
+                </td>
+                <td class="hidden w-1/3 whitespace-nowrap sm:table-cell lg:hidden">
+                    <p class="">
+                        <span v-if="item" class="inline-block w-12 opacity-70">{{ $t('domain.location.eta') }}: </span>
+                        <template v-if="item?.eta">{{ $d(item.eta, DateTimeFormat.DDD_DD_MM_hh_mm) }}</template>
+                        <span v-else-if="item" class="italic opacity-25">{{ $t('generic.no-information') }}</span>
+                    </p>
+                    <p class="">
+                        <span v-if="item" class="inline-block w-12 opacity-70">{{ $t('domain.location.etd') }}: </span>
+                        <template v-if="item?.etd">{{ $d(item.etd, DateTimeFormat.DDD_DD_MM_hh_mm) }}</template>
+                        <span v-else-if="item" class="italic opacity-25">{{ $t('generic.no-information') }}</span>
                     </p>
                 </td>
-                <td class="hidden w-1/2 whitespace-nowrap sm:table-cell">
-                    <p class="mb-2 text-sm">
-                        <span class="mr-2 inline-block w-10 opacity-50">{{ $t('domain.location.eta') }}:</span>
-                        <span v-if="item?.eta" class="font-semibold">{{ $d(item.eta, DateTimeFormat.DDD_DD_MM_hh_mm) }}</span>
-                        <span v-else>-</span>
+                <td class="hidden w-1/4 whitespace-nowrap lg:table-cell">
+                    <p class="mb-1 font-semibold">
+                        <template v-if="item?.eta">{{ $d(item.eta, DateTimeFormat.DDD_DD_MM_hh_mm) }}</template>
+                        <span v-else-if="item" class="italic opacity-25">{{ $t('generic.no-information') }}</span>
                     </p>
-                    <p class="text-sm">
-                        <span class="mr-2 inline-block w-10 opacity-50">{{ $t('domain.location.etd') }}: </span>
-                        <span v-if="item?.etd" class="font-semibold">{{ $d(item.etd, DateTimeFormat.DDD_DD_MM_hh_mm) }}</span>
-                        <span v-else>-</span>
+                    <p class="text-sm opacity-75">
+                        <template v-if="item">{{ $t('domain.location.eta') }}</template>
+                    </p>
+                </td>
+                <td class="hidden w-1/4 whitespace-nowrap lg:table-cell">
+                    <p class="mb-1 font-semibold">
+                        <template v-if="item?.etd">{{ $d(item.etd, DateTimeFormat.DDD_DD_MM_hh_mm) }}</template>
+                        <span v-else-if="item" class="italic opacity-25">{{ $t('generic.no-information') }}</span>
+                    </p>
+                    <p class="text-sm opacity-75">
+                        <template v-if="item">{{ $t('domain.location.etd') }}</template>
                     </p>
                 </td>
             </template>
@@ -52,7 +86,7 @@
                 </li>
             </template>
         </VTable>
-        <LocationEditDlg ref="editLocationDialog" />
+        <LocationEditDlg ref="editLocationDialog" :event="props.event" />
     </div>
 </template>
 <script setup lang="ts">
@@ -68,7 +102,7 @@ import { useSession } from '@/ui/composables/Session.ts';
 import LocationEditDlg from '@/ui/views/events/edit/components/LocationEditDlg.vue';
 
 interface Props {
-    event: Event;
+    event?: Event;
 }
 
 type Emit = (e: 'update:event', event: Event) => void;
@@ -82,7 +116,7 @@ const { hasPermission } = useSession();
 const editLocationDialog = ref<Dialog<Location, Location | undefined> | null>(null);
 
 async function editLocation(location: Location): Promise<void> {
-    if (hasPermission(Permission.WRITE_EVENT_DETAILS)) {
+    if (props.event && hasPermission(Permission.WRITE_EVENT_DETAILS)) {
         const editedLocation = await editLocationDialog.value?.open(location);
         if (editedLocation) {
             const updatedEvent = eventService.updateLocation(deepCopy(props.event), editedLocation);
@@ -92,21 +126,21 @@ async function editLocation(location: Location): Promise<void> {
 }
 
 async function moveLocationUp(location: Location): Promise<void> {
-    if (hasPermission(Permission.WRITE_EVENT_DETAILS)) {
+    if (props.event && hasPermission(Permission.WRITE_EVENT_DETAILS)) {
         const updatedEvent = eventService.moveLocation(deepCopy(props.event), location, -1);
         emit('update:event', updatedEvent);
     }
 }
 
 async function moveLocationDown(location: Location): Promise<void> {
-    if (hasPermission(Permission.WRITE_EVENT_DETAILS)) {
+    if (props.event && hasPermission(Permission.WRITE_EVENT_DETAILS)) {
         const updatedEvent = eventService.moveLocation(deepCopy(props.event), location, 1);
         emit('update:event', updatedEvent);
     }
 }
 
 async function updateOrders(): Promise<void> {
-    if (hasPermission(Permission.WRITE_EVENT_DETAILS)) {
+    if (props.event && hasPermission(Permission.WRITE_EVENT_DETAILS)) {
         const updatedEvent = deepCopy(props.event);
         updatedEvent.locations.forEach((location, index) => (location.order = index + 1));
         emit('update:event', updatedEvent);
@@ -114,7 +148,7 @@ async function updateOrders(): Promise<void> {
 }
 
 function deleteLocation(location: Location): void {
-    if (hasPermission(Permission.WRITE_EVENT_DETAILS)) {
+    if (props.event && hasPermission(Permission.WRITE_EVENT_DETAILS)) {
         const updatedEvent = eventService.removeLocation(deepCopy(props.event), location);
         emit('update:event', updatedEvent);
     }
